IOS-2090 Move subviews away from main class
